+++ lib/fast_ignore/rule_set_builder.rb
@@ -2,14 +2,21 @@
 
 class FastIgnore
   class RuleSetBuilder
+    # :nocov:
+    if ::FastIgnore::Backports.ruby_version_less_than?(2, 5)
+      require_relative 'backports/delete_prefix_suffix'
+      using ::FastIgnore::Backports::DeletePrefixSuffix
+    end
+    # :nocov:
+
     def self.from_args( # rubocop:disable Metrics/ParameterLists, Metrics/MethodLength
       root:,
-      ignore_rules:,
-      ignore_files:,
-      gitignore:,
-      include_rules:,
-      include_files:,
-      argv_rules:
+      ignore_rules: nil,
+      ignore_files: nil,
+      gitignore: :auto,
+      include_rules: nil,
+      include_files: nil,
+      argv_rules: nil
     )
       rule_sets = [
         from_array(ignore_rules),
@@ -28,15 +35,14 @@ class FastIgnore
     end
 
     def self.from_file(filename, project_root:, allow: false)
-      filename = ::File.expand_path(filename)
-      root = ::File.dirname(filename) + '/'
+      filename = ::File.expand_path(filename, project_root)
       raise FastIgnore::Error, "#{filename} is not within #{project_root}" unless filename.start_with?(project_root)
 
-      root = root.delete_prefix(project_root)
+      file_root = "#{::File.dirname(filename)}/".delete_prefix(project_root)
       rule_set = ::FastIgnore::RuleSet.new(allow: allow)
 
       ::IO.foreach(filename) do |rule_string|
-        parse_rules(rule_string, allow: allow, rule_set: rule_set, file_root: root)
+        parse_rules(rule_string, allow: allow, rule_set: rule_set, file_root: file_root)
       end
 
       rule_set.freeze
@@ -48,19 +54,13 @@ class FastIgnore
       end
     end
 
-    def self.from_gitignore_arg(gitignore, project_root:) # rubocop:disable Metrics/MethodLength
+    def self.from_gitignore_arg(gitignore, project_root:)
       default_path = ::File.join(project_root, '.gitignore')
       case gitignore
       when :auto
         from_file(default_path, project_root: project_root) if ::File.exist?(default_path)
       when true
         from_file(default_path, project_root: project_root)
-      when false
-        nil
-      else
-        warn 'Deprecation warning! supplying gitignore file path directly is deprecated. '\
-          'Please use gitignore: false and add your path to the ignore_files array'
-        from_file(gitignore, project_root: project_root)
       end
     end